10/5/2022
Changed the SQL server name in "ServerInfo.txt" to default load your installed SQL server vs. mine, which you'd have to change when setting up. Wanted to make that part a littler easier for everyone.

Also, I'm experimenting with tabbing to create a more organized and clean feel. But, the tabbing only looks as I intended in Visual Studio Code? So idk. I'm using that if you're curious.

Regards,
xkl