Reverted usp_GuildAcademy_withdrawMoney.
The custom procedure edit for the guild academy money withdrawal vulnerability has been reverted in both our NA2016 and CN2012 file sets: usp_GuildAcademy_withdrawMoney.

The reason is that the method I used and implemented prevented academy rewards from working correctly. Essentially, only the guild/academy master would be able to receive the rewards. This happened because I had to forcibly check if the person withdrawing money was indeed the master. This fix stopped the vulnerability but broke rewards.

Currently, the only way to properly fix this is by using a server-side client hook.